+++ tremor-pipeline/src/op/trickle.rs
@@ -16,3 +16,4 @@ pub mod operator;
 pub mod script;
 pub mod select;
 pub mod simple_select;
+pub mod window;
